Fixed all docs metadata, slugs, and tags.
The documents metadata needs to be at the top of the file, so I moved
all of the comments below the metadata. I also added "-" spacing to the
slugs. I adjusted some of the titles to better reflect the content of
the articles, I also added relevant tags and tried to add more
consistency to the tags.